Restore Discover.Query.Category 1.0.0-compatible API - SSDK-650
- See SSDK-503 for previous affecting changes
	- Commit: https://github.com/mapbox/mapbox-search-ios/commit/4e5603ed4127e67a0543d1f0666a0853ebd7808a
- Keep Maki icon updates, test updates
- Move category strings back to MapboxSearchUI
- Move Category(name, canonicalId, legacyName, icon) back to MapboxSearchUI
