--- Sources/MapboxSearch/PublicAPI/Strings.swift
@@ -1,158 +1,0 @@
-// Copyright © 2024 Mapbox. All rights reserved.
-
-import Foundation
-
-enum Strings {
-    enum Categories {
-        static let gas = NSLocalizedString(
-            "gas_station.short",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Gas title"
-        )
-        static let parking = NSLocalizedString(
-            "parking_lot",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Parking title"
-        )
-        static let food = NSLocalizedString(
-            "restaurant.alias.food",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Food title"
-        )
-        static let cafe = NSLocalizedString(
-            "cafe",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Cafe title"
-        )
-        static let restaurant = NSLocalizedString(
-            "restaurants",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Restaurants title"
-        )
-        static let bar = NSLocalizedString(
-            "bar",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Bars title"
-        )
-        static let coffeeShop = NSLocalizedString(
-            "cafe.long",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Coffee Shop title"
-        )
-        static let hotel = NSLocalizedString(
-            "hotel",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Hotel title"
-        )
-        static let gasStation = NSLocalizedString(
-            "gas_station",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Gas Station title"
-        )
-        static let chargingStation = NSLocalizedString(
-            "charging_station",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category EV Charging Station title"
-        )
-        static let busStation = NSLocalizedString(
-            "bus_station",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Bus station title"
-        )
-        static let railwayStation = NSLocalizedString(
-            "railway_station",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Train station title"
-        )
-        static let shoppingMall = NSLocalizedString(
-            "shopping_mall",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Shopping malls title"
-        )
-        static let grocery = NSLocalizedString(
-            "grocery",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Supermarket / Grocery title"
-        )
-        static let clothingStore = NSLocalizedString(
-            "clothing_store",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Clothing / shoes title"
-        )
-        static let pharmacy = NSLocalizedString(
-            "pharmacy",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Pharmacy title"
-        )
-        static let museum = NSLocalizedString(
-            "museum",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Museums title"
-        )
-        static let park = NSLocalizedString(
-            "park",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Parks title"
-        )
-        static let cinema = NSLocalizedString(
-            "cinema",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Movie Theaters title"
-        )
-        static let fitnessCentre = NSLocalizedString(
-            "fitness_center",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Gym / Fitness title"
-        )
-        static let nightclub = NSLocalizedString(
-            "nightclub",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Night clubs title"
-        )
-        static let autoRepair = NSLocalizedString(
-            "auto_repair",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Auto repair / mechanic title"
-        )
-        static let atm = NSLocalizedString(
-            "atm",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category ATM title"
-        )
-        static let hospital = NSLocalizedString(
-            "hospital",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category Hospital title"
-        )
-        static let emergencyRoom = NSLocalizedString(
-            "emergency_room",
-            tableName: "Categories",
-            bundle: .mapboxSearch,
-            comment: "Category ER title"
-        )
-    }
-}
