+++ Sources/MapboxSearchUI/Resources/en.lproj/Categories.strings
