+++ Sources/MapboxSearchUI/Strings.swift
@@ -166,6 +166,159 @@ enum Strings {
         )
     }
 
+    enum Categories {
+        static let gas = NSLocalizedString(
+            "gas_station.short",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Gas title"
+        )
+        static let parking = NSLocalizedString(
+            "parking_lot",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Parking title"
+        )
+        static let food = NSLocalizedString(
+            "restaurant.alias.food",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Food title"
+        )
+        static let cafe = NSLocalizedString(
+            "cafe",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Cafe title"
+        )
+        static let restaurant = NSLocalizedString(
+            "restaurants",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Restaurants title"
+        )
+        static let bar = NSLocalizedString(
+            "bar",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Bars title"
+        )
+        static let coffeeShop = NSLocalizedString(
+            "cafe.long",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Coffee Shop title"
+        )
+        static let hotel = NSLocalizedString(
+            "hotel",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Hotel title"
+        )
+        static let gasStation = NSLocalizedString(
+            "gas_station",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Gas Station title"
+        )
+        static let chargingStation = NSLocalizedString(
+            "charging_station",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category EV Charging Station title"
+        )
+        static let busStation = NSLocalizedString(
+            "bus_station",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Bus station title"
+        )
+        static let railwayStation = NSLocalizedString(
+            "railway_station",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Train station title"
+        )
+        static let shoppingMall = NSLocalizedString(
+            "shopping_mall",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Shopping malls title"
+        )
+        static let grocery = NSLocalizedString(
+            "grocery",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Supermarket / Grocery title"
+        )
+        static let clothingStore = NSLocalizedString(
+            "clothing_store",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Clothing / shoes title"
+        )
+        static let pharmacy = NSLocalizedString(
+            "pharmacy",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Pharmacy title"
+        )
+        static let museum = NSLocalizedString(
+            "museum",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Museums title"
+        )
+        static let park = NSLocalizedString(
+            "park",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Parks title"
+        )
+        static let cinema = NSLocalizedString(
+            "cinema",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Movie Theaters title"
+        )
+        static let fitnessCentre = NSLocalizedString(
+            "fitness_center",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Gym / Fitness title"
+        )
+        static let nightclub = NSLocalizedString(
+            "nightclub",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Night clubs title"
+        )
+        static let autoRepair = NSLocalizedString(
+            "auto_repair",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Auto repair / mechanic title"
+        )
+        static let atm = NSLocalizedString(
+            "atm",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category ATM title"
+        )
+        static let hospital = NSLocalizedString(
+            "hospital",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category Hospital title"
+        )
+        static let emergencyRoom = NSLocalizedString(
+            "emergency_room",
+            tableName: "Categories",
+            bundle: .mapboxSearchUI,
+            comment: "Category ER title"
+        )
+    }
+
     enum Feedback {
         static let report = NSLocalizedString(
             "search.feedback.reportButton.name",
